requirements/base.txt
Django==3.2.5
djangorestframework==3.12.4
django-filter==2.2.0
mysqlclient==2.0.3
django-cors-headers==3.8.0
boto3==1.12.7
requests==2.25.1
django-environ==0.4.5
django-storages==1.9.1
arrow==1.2.3
django-celery-beat==2.4.0
redis==4.4.0
celery==5.2.7
django-authtools==2.0.0